Refactor DailyWeatherPanel to use formatWeatherData function for better readability
