+++ app/utils/weatherUtils.ts
@@ -1,5 +1,5 @@
 import { format } from "date-fns";
-import { Alert } from "../models/weatherTypes";
+import { Alert, WeatherItem } from "../models/weatherTypes";
 import {
   DIRECTION_SEGMENTS,
   FAHRENHEIT_BASE,
@@ -9,8 +9,10 @@ import {
   MILES_CONVERSION_FACTOR,
   MILLISECONDS_PER_SECOND,
   MM_TO_INCHES,
+  PERCENT_MULTIPLIER,
   POPULATION_THRESHOLD,
   QUARTER_PHASE,
+  RAIN_PRECISION,
   SLICE_END_INDEX,
   THREE_QUARTER_PHASE,
   WIND_DIRECTION_DIVISOR
@@ -210,9 +212,50 @@ const formatDate = (timestamp: number) => {
   return format(new Date(timestamp * MILLISECONDS_PER_SECOND), "p 'on' PPP");
 };
 
+/**
+ * Formats the weather data for a given item.
+ * @param item - The weather item to format.
+ * @param timezone - The timezone to use for formatting time-related data.
+ * @returns An array of objects containing the formatted weather data.
+ */
+const formatWeatherData = (item: WeatherItem, timezone: string) => {
+  return [
+    { id: "Forecast", result: item.weather[0].description },
+    {
+      id: "Dew Point",
+      result: `${toCelsius(item.dew_point)}°C | ${item.dew_point}°F`
+    },
+    { id: "Clouds", result: `${item.clouds}%` },
+    { id: "Humidity", result: `${item.humidity}%` },
+    { id: "Rain", result: `${item.pop * PERCENT_MULTIPLIER}%` },
+    {
+      id: "Precipitation",
+      result:
+        item.rain !== undefined && item.rain >= 0
+          ? `${item.rain}mm | ${mmToInches(item.rain).toFixed(
+              RAIN_PRECISION
+            )}in`
+          : "0.00"
+    },
+    { id: "UV Index", result: item.uvi },
+    { id: "Moon", result: getMoonPhase(item.moon_phase) },
+    {
+      id: "Wind",
+      result: `${toKph(item.wind_speed)} kph | ${item.wind_speed} mph`
+    },
+    {
+      id: "Gust",
+      result: `${toKph(item.wind_gust)} kph | ${item.wind_gust} mph`
+    },
+    { id: "Sunrise", result: parseTime(item.sunrise, "en-US", timezone) },
+    { id: "Sunset", result: parseTime(item.sunset, "en-US", timezone) }
+  ];
+};
+
 export {
   dayOfWeek,
   formatDate,
+  formatWeatherData,
   generateAlertKey,
   getMiles,
   getMoonPhase,
